+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Add `ise_device_admin_authorization_rule` resource and data source
 - Add `ise_device_admin_authorization_exception_rule` resource and data source
 - Add `ise_device_admin_authorization_global_exception_rule` resource and data source
+- BREAKING CHANGE: Rename `profile` attribute to `profiles` of `ise_network_access_authorization_rule` resource and data source
+- Add `ise_network_access_authorization_exception_rule` resource and data source
 
 ## 0.1.5